Move prefetchData to sep file
