+++ front/app/prefetchData.ts
@@ -1,0 +1,35 @@
+import { GLOBAL_CONTEXT } from 'api/authentication/authentication_requirements/constants';
+import { fetchAuthenticationRequirements } from 'api/authentication/authentication_requirements/getAuthenticationRequirements';
+import requirementKeys from 'api/authentication/authentication_requirements/keys';
+import customPagesKeys from 'api/custom_pages/keys';
+import { fetchCustomPages } from 'api/custom_pages/useCustomPages';
+import homepageBuilderKeys from 'api/home_page_layout/keys';
+import { fetchHomepageBuilderLayout } from 'api/home_page_layout/useHomepageLayout';
+import navbarKeys from 'api/navbar/keys';
+import { fetchNavbarItems } from 'api/navbar/useNavbarItems';
+
+import { queryClient } from 'utils/cl-react-query/queryClient';
+
+const prefetchData = () => {
+  queryClient.prefetchQuery({
+    queryKey: homepageBuilderKeys.items(),
+    queryFn: fetchHomepageBuilderLayout,
+  });
+
+  queryClient.prefetchQuery({
+    queryKey: navbarKeys.list({}),
+    queryFn: () => fetchNavbarItems({}),
+  });
+
+  queryClient.prefetchQuery({
+    queryKey: customPagesKeys.lists(),
+    queryFn: fetchCustomPages,
+  });
+
+  queryClient.prefetchQuery({
+    queryKey: requirementKeys.item(GLOBAL_CONTEXT),
+    queryFn: () => fetchAuthenticationRequirements(GLOBAL_CONTEXT),
+  });
+};
+
+export default prefetchData;
